# Prohibit underscore in package names and improve package validation performance