### Context
Java 11 introduced `_` as [keyword](https://docs.oracle.com/javase/specs/jls/se12/html/jls-3.html#jls-Keyword), making it prohibited in package names.

This PR adds it to the package validation and improve the validation's performance.

Please review commit-by-commit to have better context.

----

Failure reason before the fix:
![image](https://user-images.githubusercontent.com/1841944/56750706-a3070100-6784-11e9-9312-7994a9405a87.png)

----

Speed improvements, as measure by the `JvmPackageNameBenchmark` JMH benchmark:
Before:
```
Benchmark                                                              Mode  Cnt       Score       Error  Units
JvmPackageNameBenchmark.invalid_package_name_creation                 thrpt   20   30394.430 ±   642.439  ops/s
JvmPackageNameBenchmark.package_name_with_lots_of_fragments           thrpt   20  375886.800 ± 15573.145  ops/s
JvmPackageNameBenchmark.reserved_java_keywords_package_name_creation  thrpt   20   13369.155 ±   388.070  ops/s
JvmPackageNameBenchmark.valid_package_name_creation                   thrpt   20  321294.496 ±  9347.673  ops/s
```
After:
```
Benchmark                                                              Mode  Cnt        Score       Error  Units
JvmPackageNameBenchmark.invalid_package_name_creation                 thrpt   20   864371.691 ± 65456.038  ops/s
JvmPackageNameBenchmark.package_name_with_lots_of_fragments           thrpt   20  1124642.130 ± 32982.479  ops/s
JvmPackageNameBenchmark.reserved_java_keywords_package_name_creation  thrpt   20   687804.950 ± 16936.767  ops/s
JvmPackageNameBenchmark.valid_package_name_creation                   thrpt   20   818511.320 ± 99481.866  ops/s
```

i.e. on average ~3x faster validation.